Convert all necessary template files to erbs
This commit takes all of the template files from the classifier
project that had the string `classifier` in them, and converts
them to `erb` templates that have the necessary substitutions
to inject EZBake config data.

It also adds a rake task for generating the makefile, since
the Makefile.erb is not in a directory that the packaging
framework would normally process.