Introduce per-project directory structure for configs
Prior to this commit, individual ezbake project configs were
just lein project files (with project-specific filenames) that
lived directly in the `./configs` directory.

We are going to need to support a few other files that will,
in conjuction with the lein project file, make up the full
configuration for the ezbake build.  (A specific example
is a `webserver.ini` file for a build, which will be common
to all bundled trapperkeeper artifacts, rather than being
defined by any of the upstream trapperkeeper projects.)

This commit simply changes the convention so that we now
have a `./configs/<project-name>` directory for each ezbake
project.  Inside that directory, there will be the lein
project file (named `<project-name>.clj`), plus whatever
other files we need for this build.