(PE-3516) Add functions to rename debian service files
For debhelper to correctly generate pre and post scripts for service
management, as is generally desireable, it has to also be installing the init
script. This was previously happening via the Makefile. Debian will do this
automatically if the files are placed in the debian directory with the expected
names. This commit adds functions to help accomplish this.