(PE-3516) Don't install init scripts via the makefile
The makefile template previously included logic for installing the correct init
script on different platforms. This logic is better served by moving it
entirely into the platform specific packaging. For rpm, the init script is now
installed via the spec file, and for debian it is automatically installed by
naming it correctly in the debian directory.