(PE-3516) Add status functions, implement condrestart
Having conditional restart available makes a lot of the spec file post
and pre logic cleaner. Instead of stopping the service, laying down a
file, and then starting it afterward, we can just only restart the
service if it is running, which will accomplish the same thing. This
commit adds functions for status and restart where needed and implements
the condition restart behavior described above.