(PE-3177) Add PE templates to ezbake
This commit copies the foss templates and updates them for PE. This
includes updating mocks and cows to use the correct versions for PE and
updating package dependencies to use pe-ruby and pe-java instead of
the distro packages. Pathing is also updated to use /etc/puppetlabs as
sysconfdir, /opt/puppet as the prefix, and /var/log/pe- as logdir.