(PE-3177) Add build_pe to project_data template
To switch on PE build behavior, it is necessary to add a build_pe: true
key to project_data.yaml. To do this, some templating is added to the
project_data mustache template and some logic is added to the generation
function to turn the build-type into a is-PE boolean.