(PE-3205) Take project on command line
Rather than hard-coding the project to build into stage.clj, it is
preferrable to take it as a command line argument. This commit adds
such an argument and some error handling around it. As build-target is
now encoded into the ezbake configs for the projects, we don't need to
pass it in on the command line. This commit adds some logic to pull the
value out of the ezbake config.