(maint) Don't copy dotfiles for config staging
If one were editing a config file and then ran `lein stage`, the dotfile
would be included in the copy that was turned into a project. This would
pose a problem when the Makefile template was evaluated, as it lists all
config files (including dotfiles) to be explicitly copied. The packaging
would then ignore the dotfiles and the build would fail trying to copy a
dotfile that didn't make it into the tarball. This commit prevents that
by never copying dotfiles in the configs directory into the staging
directory.