(PE-3177) Drive both build and stage from stage.clj
Previously a stub namespace had been created to automate the building of
packages. This did not support PE builds and did not accept command line
arguments for project name. This commit adds that facility to ezbake by
having the main function in stage.clj take an action and a project and
then hand that off to the correct function for staging or building. The
build function now correctly handles building PE packages, with a hard
coded PE version. The build.clj file can be removed as all of the
functionality is now within stage.clj. The lein aliases are also removed
to reflect this change, with a main added. Now, invocation is handled
using `lein run stage jvm-puppet` instead of `lein stage jvm-puppet`.