(maint) Update project_data.yaml to use template mapping
Previously functions were required to rename each of the templates that
was expected to be named using the project name. Now that the packaging
repo supports using a source/target hash in the template list, the
project_data.yaml template has been updated to use this for those
templates and the functions have been removed.