Support passing NEXUS_VERSION on Jenkins slave.
The last commit in this branch adds JVMPUPPET_NEXUS_VERSION to allow more
flexible definition of the puppetlabs/jvm-puppet dependency as well as the
jvm-puppet and pe-jvm-puppet lein project versions.

This commit supplements that behavior by generalizing the environment variable
name, adding nexus_version to project_data.yaml and the associated ezbake
generation functions, and invoking that build parameter in the uber_jar.rake
packaging task that ends up being run on a Jenkins slave such that the lein task
sees it.

Signed-off-by: Wayne <wayne@puppetlabs.com>