(PE-3652) Make JAVA_ARGS configurable
As some ezbake projects will have different needs of the jvm, this
commit allows individual projects to override the default JAVA_ARGS set
in the default template for packaging. The override can be set in the
individual project's ezbake config, using the java-args key. If no
java-args are provided, it defaults to "-Xmx192m".