(maint) Ensure expected version and source tarball match for RPM
The rpm spec templates in ezbake had realversion used in the setup macro
(which unpacks the tarball during the build) and the @rpmversion version
in the Source0 field. This worked as long as realversion == @rpmversion,
which is no longer the case when maven timestamps were introduced as
part of the version strings. The rpm version has any - replaced with .,
while the realversion does not. This commit addresses that by ensuring
that the correct version (realversion) is used in both instances.