(maint) Ensure terminus provides and obsoletes are valid rpm versions
With the introduction of maven timestamps into the version string, the
provides and obsoletes for the termini packages have invalid versions,
which is generating warnings during the build. For example:

```
warning: line 60: Invalid version (double separator '-'): 0.3.2-20140508.203329-6: Obsoletes: classifier-terminus < 0.3.2-20140508.203329-6
warning: line 61: Invalid version (double separator '-'): 0.3.2-20140508.203329-6: Provides:  classifier-terminus >= 0.3.2-20140508.203329-6
```

This commit addresses that problem but subbing . for - in all of the
terminus versions strings, which will ensure that it is a valid version
string.