(maint) Add real-name to the ezbake.rb
Because of how PE is generally laid out, the confdir for a project like
pe-classifier should be /etc/puppetlabs/classifier and the datadir
should be /opt/puppet/share/classifier. They are currently
/etc/puppetlabs/pe-classifier and /opt/puppet/share/pe-classifier. This
commit adds the real-name element to ezbake.rb, to make the non-prefixed
name available to the packaging to avoid this problem.