(EZ-15) Replace hard-coded path names.
This commit has a number of effects, including:

* Remove hard-coded path names in favor of functions that take path fragements
  and return File objects.
* Use clojure.java.io to access template resources which enables ezbake to be
  run from the git repository as a leiningen app OR as a plugin whose resources
  are stored in its jarfile.

(MAINT) Remove duplicate code.

(MAINT) Bug fix for calling ezbake on directory.

Previously when passing a directory to ezbake in place of a project name, the
path to the config directory within that project directory was not calculated
correctly. This commit fixes that bug by delegating the creation of a File
object representing that directory to its own function that tests whether the
File object isDirectory and if not falls back to treating the project parameter
as a project name and creating the directory based on the old approach.