(PE-5349) Add extra install options
This commit adds support for upstream projects to supply arbitrary
install commands to be run in addition to the standard steps. They are
supplied on a per platform basis similarly to the preinst and
dependencies. This commit also adds them to ezbake.rb so they can be
referenced in the rules.erb and ezbake.spec.erb for both foss and pe.