(PE-3002) Fix bug if no bin script
The changes related to supporting CLI apps in ezbake projects
were intended to only create a wrapper bin script if the project
actually provided at least one CLI tool.  However, in the redhat
spec, we were trying to install the bin script even if that
condition hadn't been met.  This commit should fix that issue,
and allow RPMs to be built for projects that don't have any cli
apps.