(PE-5792) Create /var/log/$project during packaging
Previously the init script was used to create and ensure correct
permissions on the log dir during service start. This doesn't work for
el7, where systemd is in play. systemd has no facility to create a
logdir, and the user that the application will be running as will not be
able to make a directory in /var/log, so the packaging has to do it.
Once the directory has been created, logback will be able to create the
log file in the directory. This commit adds the logdir to the deb and
rpm packaging and ensures that the directory is 700 and owned by the
user and group for the ezbake project.