(maint) Refactor get-* functions into one function
Previously several functions were present to get a value from ezbake
configs, one for each key in the config. This led to a lot of repetition
and new code being needed to get new keys when they were added. This
commit refactors those functions into one single function that takes a
keyword to lookup as an extra parameter. It also updates the calls to
use the new function and removes the now unneeded old functions.