(PE-5756) Add post-start action to ezbake
This adds the post-start-action to the ezbake.rb so that it will be
available to be used with packaging. This commit also includes changes
to the foss and PE packaging to use the post-start-action if it is
specified. It is implemented so that it only is invoked if the start was
successful and it does not impact the exit status of the service start
call.