(PE-5757) Add support for additional binaries to ezbake
Projects built using ezbake may need to supply their own binaries. This
commit allows them to do so by putting them in 'resources/ext/bin'. It
Adds sections to the global makefile to install the files correctly and
to the specfile templates to add the files to the %files section.