(SERVER-3) Update EL init scripts to invoke the daemon from INSTALL_DIR
On el systems, if the service is invoked using init.d directly, while in
a root owned directory, it will fail to run several external commands
during startup. This isn't a problem for debian because we pass in a
directory to run from that will be accessible to the service user. This
commit implements the same behavior for EL and sles. On el this is
accomplished by using pushd (which means the shell invoking the init
script must be bash) to move into the $INSTALL_DIR directory, which
defaults to the home directory of the service user. On sles, the -c
option is passed to startproc (which means it must also be passed to
checkproc and killproc), which will ensure the daemon is started from
the $INSTALL_DIR directory.