(maint) Use pushd/popd instead of -c for sles init scripts
-c did not behave as expected. Sles used the directory specified as a
chroot and then tried to find the daemon (/opt/puppet/bin/java) within
the directory, where it wouldn't exist. This commit updates the proc
calls to remove the -c, and replaces the one in the start function with
a pushd/popd. The use of pushd means that the shell has to be bash, so
that is also updated.