(SERVER-8) Remove ExecStop, add KillMode to systemd services
The currently specified ExecStop is already the default behavior in
systemd, so it is not required. It also causes issues if the service
hasn't started and there is no MAINPID defined. This commit removes that
setting and adds KillMode instead. KillMode=process, as described in
http://www.freedesktop.org/software/systemd/man/systemd.kill.html, will
only target the main process, which should be all that exists in our
clojure applications.