(SERVER-24) Moved env dir code
Previously, a seperate script (cli-env) was run before running the main
CLI script for an EZ Bake project. This ended up providing us with a bit
of a chicken/egg problem for where to look for subcommands and where the
system's Java executable is located.

This commit removes the cli-app script, and directly places the logic at
the top of the project's main CLI script.