(EZ-1) Update sysconfig name file to not be replaced on install
This commit adds a `%config(noreplace)` directive to the sysconfig name
file in the RedHat spec erb files.  This allows an upgrade package being
installed to not replace any customizations that a user might have done
to the name file after a previous install.

This commit also has one other minor update to the link to the
"Developing ezbake" section in the `README.md`.