(MAINT) Add explicit nrepl references
There appears to be a bug in leiningen where it will exclude
tools.nrepl from an uberjar if it is a transitive dependency of
your project (as opposed to a first-class dependency listed in
the project.clj file that you are building the uberjar from).

See https://github.com/technomancy/leiningen/issues/1762 .

This commit adds an explicit dependency to all of our ezbake
project config files, to make sure that we are including the
nrepl code in the uberjars.