(PE-6711) Unify logging configs for pe services with the pe-module
This commit updates the logging configs for pe-puppetserver and
pe-console-services to be the configs we intend to use in production so
that we don't need to manage them using puppet or lay down new ones in
the installer. This way when a user need not wait for puppet to run in
order to see logging changes for pe services.