(EZ-21) Lay down logrotate for ezbake packages
Ezbake packages can generate a lot of logs, so this commit adds
automatic log rotation for anything in the /var/log/$project directory.
Unfortunately, logrotate scripts are terrible because between certain
versions a line is required in the configuration, without which errors
are generated, but that is not recognized on earlier versions of
logrotate. There are two logrotate configs here, the legacy one is
designed for older rpm and debian systems, while the standard config is
for modern platforms.