(EZ-15) Fix lein-ezbake rebased onto TK-112 changes.
Previous to this commit in the `lein-ezbake` work there were several functions
with redundant behavior factored into the single function called
`cp-shared-files`. TK-112 modified one of the removed functions by causing it to
copy files from the "global" templates but in fact the entire "global" template
directory is already copied by another function earlier during the staging
sequence.

The real useful work of TK-112 was in causing the modified function to return
the additionally-copied cli-app files so their names can be used later on by
config generation function. During rebase this function was removed entirely by
mistake but that's okay. Instead of keeping that function around this commit
simply lists the contents of the appropriate staging directory and relativizes
the strings for later use during config generation.