(EZ-15) Create uberjar during `stage` action.
This is prevents us from having to run the uberjar task within rake or make or
whatever, so then we don't have to worry about passing along the correct
leiningen profile at those points. It's actually pretty difficult to determine
which profiles are in use from within a lein task.