(PE-7078) Add 'start-after' parameter to ezbake
Prior to this commit, ezbake projects
using systemd are not able to specify a
startup dependency on other ezbake services.

Because of this, some ezbake services can
fail to start on reboot.

This commit adds a 'start_after' parameter to allow
specification of a systemd service dependency
in a project's project.clj.