(PE-7614) Fix SLES-11 pe-puppetserver init script
Without this patch the pe-puppetserver service in PE fails to start on SLES-11.
Evidence of the root cause emerges when using the status method of the service
after the failure occurs:

    $ sudo /etc/init.d/pe-puppetserver status
    checkproc: Can not handle pid file /var/run/pe-puppetserver/pe-puppetserver.pid with pid `'
                                                   dead

This error is caused by $pid being unbound, presumably because the java process
never starts.  The script prints the unbound $pid variable with a trailing
newline to $PIDFILE.  This newline caused all of the `[ -s "${PIDFILE}" ]`
guard checks to pass when they should have failed because the -s test returns
true when the file has a non-zero size.

In addition to this error, there are other issues that prevent the script from
parsing correctly.  For example, making a surgical fix with focus on only
$retval, and $pid the following is still returned by `sudo
/etc/init.d/pe-puppetserver start`:

    /etc/init.d/pe-puppetserver: line 63: [: missing `]'

And the following is returned by `sudo /etc/init.d/pe-puppetserver stop`:

    /etc/init.d/pe-puppetserver: line 75: [: : integer expression expected

Given this cluster of issues, this patch resolves the problem by applying the
following changes.

Make sure $retval is bound immediately after calls to `rc_status`
because the subsequent code makes assumptions that it is bound and
contains an integer value.

Fail early when the return value of `rc_status` is non-zero because the
PID likely never started or died quickly.  Failing early prevents other
guard checks from executing unpredictably and focuses on the concern of
starting and checking the process via the PID.  If the script does not
fail early, errors and warnings as the script proceeds unpredictably
make troubleshooting more difficult.

More minor changes, but relevant to the errors encountered:

In general, replace `[ "$retval" -eq 0 ]` with `[ "$retval" = 0 ]`
because the latter behaves as expected when $retval is unbound or
contains a non-integer value.  The use of `-eq` with a string is less
robust because it returns `integer expression expected` for strings that
don't look like integers.

In general, replace `echo "${pid}" > "${PIDFILE}"` with `echo -n "${pid}" ...`
in an effort to be more robust with the pattern of using `[ -s "${PIDFILE}" ]`
because `echo` without `-n` will always send at least one byte, a newline, to
the file and `-s` tests against a non-zero size, which will never happen
without the `-n`.

In general, quote all bare variables passed as positional arguments because the
size of the argument vector changes if the variable is unbound without the
quotes.  This change is more robust when using fixed positional arguments as is
the case with functions like `wait_for_app`.