(SERVER-338) Add net-tools dependency for service unit files
Without this patch the init scripts and unit files make calls to
netstat, which may or may not exist on the system.  This is a problem
because the service will fail to start if netstat is not present.  This
patch addresses the problem by expressing net-tools as a dependency of
the packages which will ensure netstat is present prior to service
management.