(SERVER-344) Use runuser instead of su with foreground
Without this patch the `puppetserver foreground` subcommand does not
work out of the box on our EL-7 systems.  The command prompts for a
password because the puppet user does not have a valid shell.  There is
no known work-around without invasive modification to the system
configuration, e.g. giving puppet a valid shell or adjusting the PAM
configuration.

This patch addresses the problem by switching `su` to `runuser` which is
specifically designed to execute commands in shells as users with
otherwise invalid login shells.