Version 0.3.0-SNAPSHOT
Without this patch the master branch of ezbake has the same version as
the stable branch.  This is a problem because we're going to need to
make changes to the master branch for the new AIO path specification,
but we also need to do releases of 0.2.x for previous puppet-server
stable releases and such.

This patch addresses the problem by establishing the master branch as
version 0.3.0-SNAPSHOT while the stable branch will remain on 0.2.x.