(SERVER-344) Choose best method to become puppet
Previously, only the su command was used to switch to the puppet user
and launch Puppet server in the foreground. This does not work on some
systems using SELinux. This commit attempts to use the runuser command
instead if it is installed and the user is running as root, otherwise
sudo or su is used.