(SERVER-369) Move /var/log/{,puppetlabs}/puppetserver for EL-7
Without this patch the log directory for puppetserver is incorrect and
does not match the proposed [Puppet Installation Layout][proposal].  The
current specified [layout] does not account for the log directory of
puppetserver, hence the proposal.

This patch addresses the problem by changing the log location, the
location of Java heap dumps, and the logrotate configuration to reflect
the newly proposed log directory.

[proposal]: https://github.com/puppetlabs/puppet-specifications/pull/27/files
[layout]: https://github.com/jeffmccune/puppet-specifications/blob/7880849a739ab7d16555fb58c53d10c788f4b958/file_paths.md