(SERVER-508) Use su before sudo
Without this patch the sudo command is used to start puppetserver when
EUID is 0.  This is a problem on Ubuntu 14.04 because the puppet user is
not in the sudoers file and the sudo command simply exits without
executing anything.

This patch addresses the problem by trying su before trying sudo.  sudo
is used as a last resort.