(SERVER-502) Add puppetserver daemon subcommand
Without this patch the puppetserver java process does not start on
Ubuntu 14.04.  This patch addresses the problem by moving the java
command and the delicate shell escaping of the java command line
arguments into a specialized subcommand bash script.  This makes it
easier to handle the argument vector while also capturing STDOUT and
STDERR to daemon log files in a way that is compatible with
start-stop-daemon