(SERVER-508) Run puppetserver foreground directly when correct user
Without this patch the puppetserver foreground command runs the
puppetserver command through `su` when the parent context is running as
non-root.  This is a problem because it's not necessary to switch users
in all cases.  If we're already the correct user we can run the command
directly, and if we're not the correct user running through `su` is
likely the wrong thing to do because this will prompt for a password
when non-root and the correct user to run as likely has no password set.

This address Jeremy's comment to that effect at [1].

[1]: https://github.com/puppetlabs/ezbake/pull/212#issuecomment-85141407