(maint) Update HeapDumpPath to AIO spec
Without this patch the HeapDumpPath is set to /var/log/<name>, e.g.
/var/log/puppetserver.  This is a problem because this directory does
not exist according to the [Puppet Installation Layout][layout] and as
such the heap dump file may never be written because the parent
directory does not exist.

This patch addresses the problem by updating the HeapDumpPath to match
the log directory of the layout spec.

[layout]:
https://github.com/puppetlabs/puppet-specifications/blob/2abde272af859ffb434cb62725e6f54ca65da704/file_paths.md