(SERVER-502) Fix <project>-daemon.log directory location
Without this patch the service fails to start on all debian based
platforms.  The root cause is that bash, started by start-stop-daemon as
a non-root user, cannot write to the parent directory of the daemon log
file and as a result the service process itself is never started.  The
error message is silent making this difficult to troubleshoot.

This patch addresses the problem by writing the daemon log file to the
correct directory mentioned in the [Puppet Installation Layout][layout].
This directory has the correct ownership and permissions to allow the
process to create and write to the file.

[layout]: https://github.com/puppetlabs/puppet-specifications/blob/2abde272af859ffb434cb62725e6f54ca65da704/file_paths.md