(RE-3976) Add 'Should-Start' headers to SUSE init script template
Prior to this commit, the pe-console-services init script on SLES 11
does not contain the LSB headers for dependency-based service ordering.
As a result, on SLES 11, the PE console is sometimes unavailable after
a reboot.

This commit updates the SUSE init script template to include the templated
'Should-Start' LSB headers.

For consistency, we make the same change to the EL init script template.