(maint) Use templated START_TIMEOUT in Debian init scripts
Prior to this commit, the START_TIMEOUT in the ezbake
FOSS and PE RedHat and SUSE init script is templated,
but is hardcoded for Debian.

When the template value is evaluated for RedHat and SUSE,
the START_TIMEOUT is 180.

This commit updates the Debian init script to use the
templated value for START_TIMEOUT.