(EZ-38) Update install.sh to default to using an inside voice
Previously install.sh would be very chatty and tell users lots of things
they might not want to know during package installation. This commit
changes the default behavior to be quiet. This behavior can be toggled
using the EZ_VERBOSE environment variable. If set, the same previous
verbosity is enabled.