(EZ-38) Update packaging to enable verbose install.sh
With the new default of quiet install.sh, we still want to be able to
easily track down packaging problems, so this commit updates the package
to set the EZ_VERBOSE to 1 in the debian and redhat packaging to ensure
that all details are printed during package build. Pre and post scripts
do not have the environment variable set, so they remain quieted.

Conflicts:
	resources/puppetlabs/lein-ezbake/template/foss/ext/redhat/ezbake.spec.erb