(RE-4396) Do not build stable or testing
We should only be building debian packages for the codenames of the
different available debian platforms. We are currenly building packages
twice for each platform as the moment, since we are building both wheezy
and stable, when we could just be building wheezy. From this logic, we
should not be building packages for oldstable, stable, testing, or
unstable. Since stable and testing are currently the only ones in that
list here, this commit removes them.