(EZ-44) Set sudo HOME for foreground
This commit changes executions of commands via sudo from the foreground
subcommand to be done with a -H option.  The -H option causes the HOME
environment variable for the process running in sudo to be set to that
of the default home directory for the target user.  The -H overrides the
default behavior on Ubuntu, which is to preserve the HOME variable value
for the user which ran the foreground subcommand.  This change makes
command executions running via sudo consistent with those done for su and
runuser.