(PDB-1455) Correct termini install.sh rubylibdir fallback autodetection
The autodetection routine for AIO pathed Puppet rubylibdir wasn't working.
This showed up when we tried to point our PuppetDB source based tests at a
the ezbake upgrade branch.

This patch adds the Puppet 4/AIO path to PATH when trying to call ruby to
figure out where the terminus belongs.

Signed-off-by: Ken Barber <ken@bob.sh>