(PDB-1455) Correct AIO rubylibdir
Previous commit was incorrect in that it relied on the wrong config category
for the proper path (vendorlibdir instead of vendordir), this commit sets its
right.

It also fixes the foss termini spec rubylibdir to use the same mechanism as
install.sh.

The debian usage was different also for FOSS, so I've changed them both to use
the same method as install.sh for consistency.

Signed-off-by: Ken Barber <ken@bob.sh>