(re-4845) Upgrade from pre-aio to aio paths fails to stop service
Because of the path changes and the change to the name of the jar file
when upgrading from, for example, puppetdb 2.3.x to 3.0.x, the new
init script that gets installed is not able to stop the old service.

This updates the search constraints for the PID of the process to look for
the name of the $JARFILE, which should allow the init script to stop and
restart during a package upgrade. There is a related change to puppetdb
to change the name of the jarfile from 'puppetdb-release.jar' to
'puppetdb.jar' to match.