(PDB-1455) Make permissions to data dir less permissive
This patch removes world read-ability to the data directory. For PuppetDB at
least we want to use this directory to replace the old create_varlibdir
functionality instead, but the directory is too permissive for storing HSQLDB
and KahaDB data.

This patch modifies the chmod and permissions of that directory to 770 instead
of 775. It also modifies the Redhat specs to use that permission.

Signed-off-by: Ken Barber <ken@bob.sh>