(maint) Don't pass hardcoded PE_VER to rake uber_build task
Previously, we were hardcoding "PE_VER" as 3.7 and passing that along when
executing "rake pe:jenkins:uber_build." This commit removes that hardcoding,
which means a user can set the environment variable PE_VER and have it
respected by the rake task. If not provided, the rake task will error out.