(RE-4779) Restart service on upgrade
In the postinst script, we have the potential for making changes to the
service on upgrade. Since we're already restarting the service for rpm,
we should be doing the same thing for deb. However, we want to ensure
this happens after any changes in the postinst script happen.

This is for both PE and foss, becuase, why not?