(PE-11368) Use real_name in logrotate files
In the new FS, both pe-puppet-server and puppet-server put their logs in
/var/log/puppetlabs/puppet-server (or console-services or puppetdb).
Previously the logrotate files referenced project in the directory
instead of real name, which would lead to logrotate trying to rotate a
non-existent directory such as /var/log/puppetlabs/pe-puppet-server.
This commit updates the templates to use real_name instead to avoid
this.