(EZ-52) Add custom created directories to the rpm
Previously in ezbake, custom directories would be added to the buildroot
for rpm, but would not be added to the %files section of the rpm, which
means they would not be included in the package generated. This commit
adds %dir declarations for each dir in create_dirs, and assigns it the
same permissions and ownership that the install.sh postinst will.