(RE-6278) Update permissions for "projconfdir".
Prior to this commit, Ezbake packages install the
'projconfdir' with 'root:root' ownership and
'755' permissions.

We should probably remove read/execute permissions
to this directory for the 'other/world' group, but
preserve the same permissions for project service
accounts.

This commit updates Ezbake to (1) install the
'projconfdir' with '750' permissions and (2) set
user/group ownership to a project's specific
user and group.