(maint) Do not use virtual package for java dep
Prior to this commit, we were requiring virtual packages to satisfy java
runtime dependencies on deb based platforms. This was fine until we get
to Ubuntu 16.04 Xenial. On Xenial, Java7, Java8, and Java9 all provide
java8-runtime-headless. This is a problem, because it means that java9
is automatically installed to satisfy the specified runtime requirement
of java8-runtime-headless. We're not ready to update puppetserver with
Java9 support yet.

This commit switches from the virtual packages to the actual java
package that we want installed.