(EZ-72) Support split bootstrap in configuration
This commit adds support to ezbake for splitting a project's bootstrap files
between /etc/puppetlabs/{project}/services.d/ and
/opt/puppetlabs/server/apps/{project}/config/services.d/

This serves the ultimate goal of putting user-configurable service entries
in one location, and the services that should not be modified in another.

It accomplishes this by adding two new settings to ezbake which can be
specified in the :lein-ezbake section of a project.clj
* :split-bootstraps - A boolean to cause ezbake to tell trapperkeeper that
  it should look for it's bootstrap files in the two directories mentioned
  above.
* :private-config-dir - A directory in the project, who's contents will be
  layed down in the /opt directory mentioned above.

This change is backward compatible. If :split-bootstraps is false, ezbake
will continue to work the same as before, expecting a "bootstrap.cfg" in
/etc/puppetlabs/{project}/bootstrap.cfg.