(EZ-48) Add support for debian systemd
This commit updates both the PE and foss build pipelines to correctly
install systemd scripts for packages built with EZBake. It also cleans
up some of the logic in the {pre,post}{inst,rm} debian scripts.

This only upgrades Ububtu Xenial to systemd, even though it really
should include Debian Jessie. We disable the service in the prerm, which
prevents us from recording and preserving the state of the service on
upgrade. This is addressed in this change, so it should be more easily
fixed for future upgrades for Jessie.