(RE-7420) Build FOSS ezbake projects on sles
This unfortunately will not work for builds outside of the puppetlabs
ecosystem. There are no publicly available sles mirrors available, so we
have created our own in house mirror. This is not available to the
public. The sles mocks created with the rpmbuilder module are configured
to hit this internal repo, but to not fail if the user doesn't have
access to it.