(SERVER-336) Add boolean to disable logrotate
This commit adds a new boolean ezbake setting to disable logrotate in a project

SERVER-366 covers work to limit the amount of disk space puppetserver will use
for its logs. We decided to accomplish that by utilizing the built in features
of the logback library. Logback will:
* Do daily rotating of logs
* Rotate the logs out once they reach a certain size
* Limit the total size of all the puppetserver log file

Logrotate is no longer needed, so we needed a way to disable logrotate in ezbake.
The change in this commit is backwards compatible, as it defaults to turning
logrotate on.