(EZ-97) Kill background Java process if start script terminated
This commit will trap HUP, INT, and TERM signals sent to the start
script.  If any is trapped, the associated Java process that the start
script spawned will be terminated.  This is done so that if a 'restart'
is issued while a start is already in process, the Java process for the
first start attempt will be cleaned up before the second start is done.
Previously, without this cleanup, the second start, depending upon
timing, might not happen at all and an erroneous exit code for the
'restart' might be returned.