(EZ-97) Add missing check for pid status after kill in aborted start
This commit adds one extra `kill -0 <pid>` check to the logic in the
start script which confirms whether an attempt to kill the Java process
for an aborted start was successful.  Without this check, an erroneous
"Java process not killed" trace message might be written even if the
process actually was killed.