(EZ-98) Fail an ezbake package build attempt if restart-file not set
This commit causes an ezbake package build attempt to fail if the
`restart-file` setting is not specified in the ezbake config and/or
global section of the package's Trapperkeeper config.  This check is
being added because `restart-file` is essentially a required setting
for Ezbake 1.0+ in that it would be impossible to determine when a
service has been completely started without the use of the setting.