(EZ-100) Move first restart-file counter check before HUP in reload
This commit moves the check for the initial value of the restart-file
counter to be before performing a HUP in the reload script.  Previously,
the initial check was done after the HUP was performed, leading to the
possibility that the service may have already been reloaded (and counter
incremented) before any counter incrementation checks are done.  This
race condition previously could have caused the reload script to hang
until timing out and failing.