(EZ-101) Add failure example for old TK version w/o restart file support
This commit adds an example to the `README.md` of the startup failure
message that would be seen if a package were built using a TK version
which does not support the `--restart-file` command line option that
EzBake now passes into the Java command line for a service.

The README.md doc also includes a link to the `restart-file` feature
documentation in the Trapperkeeper repo.